Moved env handler and personality into frameworks.
